app/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
gunicorn==20.1.0
